Refactor types and add hook support
