+++ package.json
@@ -120,6 +120,7 @@
     "fs-extra": "^11.1.1",
     "get-port": "^7.0.0",
     "giget": "^1.1.3",
+    "hookable": "^5.5.3",
     "immer": "^10.0.3",
     "is-wsl": "^3.0.0",
     "jiti": "^1.21.0",
