+++ src/index.ts
@@ -2,5 +2,5 @@
  * @module wxt
  */
 export * from '~/core';
-export * from '~/types/external';
+export * from '~/types';
 export * from '~/version';
